build: improve AddCMockaTest `LINK_OPTIONS` docs
The `LINK_OPTIONS` argument for `add_cmocka_test()` doesn't
set the `LINK_OPTIONS` CMake property, as it was only added
in CMake 3.13.

These options are mostly the same, but some options that are valid
`LINK_OPTIONS` are not valid `LINK_FLAGS` (e.g using `LINKER:` prefixes)

This commit documents that the `LINK_OPTIONS` argument instead
is used to set the `LINK_FLAGS` CMake property.

Reviewed-by: Andreas Schneider <asn@cryptomilk.org>