tests: Add test for test_calloc()
calloc() is prone to integer overflow on multiplication of its arguments.
glibc, musl, and uclibc test for that in its implementations and return NULL in
that case with errno set to ENOMEM.  cmocka lacks such a check and passes all
kinds of overflown values to test_malloc() with different ways to fail.

Signed-off-by: Alexander Dahl <ada@thorsis.com>
Reviewed-by: Andreas Schneider <asn@cryptomilk.org>