Allow macros as input for convenience macros
Nested macros like the cmocka macros that make use of # or ## are not expanded.
The `cmocka_tostring` macro makes sure that you can use for example macros
redefining symbols like functions.

When using cmocka on embedded devices, it may happen that the original function
is necessary for it to boot correctly as well as a mocked version for unittests
with a prefix `unittest_`. To simplify this, the compiler can be given
definitions via the `-D` argument. These replacements are now also possible for
the invocations of the cmocka convenience macros like "will_return",
"expect_value", ...

Reviewed-by: Andreas Schneider <asn@cryptomilk.org>