feat(output-overrides): refactor to use structure parameter
Review suggested to generalise and future-proof the work to allow any
number of callbacks to be set in the future. This can be achieved by
simply adding additional callbacks to the CMCallbacks structure.

Any unspecified callbacks are reset to the default behaviour. This
ensures that old code can be recompiled against a newer implementation of
cmocka with no code changes required.

Signed-off-by: Micha WERLE <micha@michaelwerle.com>
Reviewed-by: Andreas Schneider <asn@cryptomilk.org>